+++ app/Http/Controllers/IndexController.php
@@ -2,13 +2,16 @@
 
 namespace Corp\Http\Controllers;
 
+use Corp\Repositories\SlidersRepository;
 use Illuminate\Http\Request;
 
 class IndexController extends SiteController
 {
-    public function __construct()
+    public function __construct(SlidersRepository $s_rep)
     {
         parent::__construct(new \Corp\Repositories\MenusRepository(new \Corp\Menu));
+
+        $this->s_rep = $s_rep;
 
         $this->bar = 'right';
         $this->template = config('settings.THEME').'.index';
@@ -21,8 +24,28 @@ class IndexController extends SiteController
      */
     public function index()
     {
+        $sliderItems = $this->getSliders();
+
+        $sliders = view(config('settings.THEME').'.slider')->with('sliders',$sliderItems)->render();
+        $this->vars = array_add($this->vars,'sliders',$sliders);
 
         return $this->renderOutput();
+    }
+
+    public function getSliders()
+    {
+        $sliders = $this->s_rep->get();
+
+        if($sliders->isEmpty()) {
+            return FALSE;
+        }
+
+        $sliders->transform(function ($item, $key) {
+            $item->img = config('settings.slider_path').'/'.$item->img;
+            return $item;
+        });
+
+        return $sliders;
     }
 
     /**
