+++ app/Providers/AppServiceProvider.php
@@ -3,6 +3,7 @@
 namespace Corp\Providers;
 
 use Illuminate\Support\ServiceProvider;
+use Illuminate\Support\Facades\Blade;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -13,7 +14,10 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        //
+        Blade::directive('set', function ($exp) {
+            list($name,$val) = explode(',',$exp);
+            return "<?php $name = $val ?>";
+        });
     }
 
     /**
