+++ app/Repositories/SlidersRepository.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Corp\Repositories;
+use Corp\Slider;
+
+
+class SlidersRepository extends Repository
+{
+    public function __construct(Slider $slider)
+    {
+        $this->model = $slider;
+    }
+}
