+++ config/settings.php
@@ -4,6 +4,7 @@ return [
 
     'THEME' => env('THEME', 'pink'),
 
+    'slider_path' => 'slider-cycle'
 
 ];
