+++ resources/views/pink/index.blade.php
@@ -3,3 +3,7 @@
 @section('navigation')
     {!! $navigation !!}
     @endsection
+
+@section('slider')
+    {!! $sliders !!}
+    @endsection
